Fix volume for flite
The backend stored the volume as int, while the API uses double between
0 and 1, make it consistent, so the getter will return sensible values.

+++ src/plugins/tts/common/qtexttospeechprocessor_p.h
@@ -119,7 +119,7 @@ private:
     volatile bool m_paused;
     double m_rate;
     double m_pitch;
-    int m_volume;
+    double m_volume;
     QSemaphore m_speakSem;
     QString m_nextText;
     int m_nextVoice;
